+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 - Error handling - Display the right errors on create/update.
 - Value Chart - Fix previous period count regression due to filterType introduction.
+- Smart field - Enable search on smart fields.
 
 ## RELEASE 1.3.44 - 2016-10-11
 ### Fixed